// src/app/shared/services/model/model.ts
export interface LoginCredentials {
  userName: string;
  password: string;
  ipAddress: string;
  browser: string;
}


export interface UserData {
  userName: string;
  ipAddress: string;
  browser: string;
  loggedUserId: string;
  loggedUserEmail: string;
  loggedUserType: string;
  fullName: string;
  active: string;
  status: string
}


export interface User {
  userName: string;
  mobile: string;
  email: string;
  password: string;
  confirmPassword: string;
  otp: string;
  userType:string;
}